+++ doc/bug_report_template.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-# Moved to .github/ISSUE_TEMPLATE/bug_report.md
+# Moved to .github/ISSUE_TEMPLATE/bug-report.md
 # Please update your bookmarks