+++ spec/dummy_app/app/controllers/test_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class TestController < ActionController::Base
+class TestController < ApplicationController
   def user_for_paper_trail
     Thread.current.object_id
   end