+++ spec/models/article_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates an ignored column" do
+  context "when updating an ignored column" do
     it "not change the number of versions" do
       article = described_class.create
       article.update(title: "My first title")
@@ -19,7 +19,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates an ignored column with truly Proc" do
+  context "when updating an ignored column with truly Proc" do
     it "not change the number of versions" do
       article = described_class.create
       article.update(abstract: "ignore abstract")
@@ -27,7 +27,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates an ignored column with falsy Proc" do
+  context "when updating an ignored column with falsy Proc" do
     it "change the number of versions" do
       article = described_class.create
       article.update(abstract: "do not ignore abstract!")
@@ -35,7 +35,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates an ignored column, ignored with truly Proc and a selected column" do
+  context "when updating an ignored column, ignored with truly Proc and a selected column" do
     it "change the number of versions" do
       article = described_class.create
       article.update(
@@ -59,7 +59,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates an ignored column, ignored with falsy Proc and a selected column" do
+  context "when updating an ignored column, ignored with falsy Proc and a selected column" do
     it "change the number of versions" do
       article = described_class.create
       article.update(
@@ -86,7 +86,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates a selected column" do
+  context "when updating a selected column" do
     it "change the number of versions" do
       article = described_class.create
       article.update(content: "Some text here.")
@@ -95,7 +95,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates a non-ignored and non-selected column" do
+  context "when updating a non-ignored and non-selected column" do
     it "not change the number of versions" do
       article = described_class.create
       article.update(abstract: "Other abstract")
@@ -103,7 +103,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates a skipped column" do
+  context "when updating a skipped column" do
     it "not change the number of versions" do
       article = described_class.create
       article.update(file_upload: "Your data goes here")
@@ -111,7 +111,7 @@ RSpec.describe Article, type: :model, versioning: true do
     end
   end
 
-  context "which updates a skipped column and a selected column" do
+  context "when updating a skipped column and a selected column" do
     it "change the number of versions" do
       article = described_class.create
       article.update(
@@ -141,7 +141,7 @@ RSpec.describe Article, type: :model, versioning: true do
       ).to(eq("content" => [nil, "Some text here."]))
     end
 
-    context "and when updated again" do
+    context "when updated again" do
       it "have removed the skipped attributes when saving the previous version" do
         article = described_class.create
         article.update(
@@ -183,6 +183,81 @@ RSpec.describe Article, type: :model, versioning: true do
       expect(PaperTrail::Version.count).to(eq(2))
       expect(article.versions.size).to(eq(2))
       expect(article.versions.map(&:event)).to(match_array(%w[create destroy]))
+    end
+  end
+
+  context "with an item" do
+    let(:article) { Article.new(title: initial_title) }
+    let(:initial_title) { "Foobar" }
+
+    context "when it is created" do
+      before { article.save }
+
+      it "store fixed meta data" do
+        expect(article.versions.last.answer).to(eq(42))
+      end
+
+      it "store dynamic meta data which is independent of the item" do
+        expect(article.versions.last.question).to(eq("31 + 11 = 42"))
+      end
+
+      it "store dynamic meta data which depends on the item" do
+        expect(article.versions.last.article_id).to(eq(article.id))
+      end
+
+      it "store dynamic meta data based on a method of the item" do
+        expect(article.versions.last.action).to(eq(article.action_data_provider_method))
+      end
+
+      it "store dynamic meta data based on an attribute of the item at creation" do
+        expect(article.versions.last.title).to(eq(initial_title))
+      end
+    end
+
+    context "when it is created, then updated" do
+      before do
+        article.save
+        article.update!(content: "Better text.", title: "Rhubarb")
+      end
+
+      it "store fixed meta data" do
+        expect(article.versions.last.answer).to(eq(42))
+      end
+
+      it "store dynamic meta data which is independent of the item" do
+        expect(article.versions.last.question).to(eq("31 + 11 = 42"))
+      end
+
+      it "store dynamic meta data which depends on the item" do
+        expect(article.versions.last.article_id).to(eq(article.id))
+      end
+
+      it "store dynamic meta data based on an attribute of the item prior to the update" do
+        expect(article.versions.last.title).to(eq(initial_title))
+      end
+    end
+
+    context "when it is created, then destroyed" do
+      before do
+        article.save
+        article.destroy
+      end
+
+      it "store fixed metadata" do
+        expect(article.versions.last.answer).to(eq(42))
+      end
+
+      it "store dynamic metadata which is independent of the item" do
+        expect(article.versions.last.question).to(eq("31 + 11 = 42"))
+      end
+
+      it "store dynamic metadata which depends on the item" do
+        expect(article.versions.last.article_id).to(eq(article.id))
+      end
+
+      it "store dynamic metadata based on attribute of item prior to destruction" do
+        expect(article.versions.last.title).to(eq(initial_title))
+      end
     end
   end
 end