+++ spec/models/gadget_spec.rb
@@ -12,13 +12,13 @@ RSpec.describe Gadget, type: :model do
       expect { gadget.update_attribute(:name, "Hammer") }.to(change { gadget.versions.size }.by(1))
     end
 
-    context "ignored via symbol" do
+    context "when ignored via symbol" do
       it "doesn't generate a version" do
         expect { gadget.update_attribute(:brand, "Picard") }.not_to(change { gadget.versions.size })
       end
     end
 
-    context "ignored via Hash" do
+    context "when ignored via Hash" do
       it "generates a version when the ignored attribute isn't true" do
         expect { gadget.update_attribute(:color, "Blue") }.to(change { gadget.versions.size }.by(1))
         expect(gadget.versions.last.changeset.keys).to eq %w[color updated_at]