+++ spec/models/skipper_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe Skipper, type: :model, versioning: true do
   it { is_expected.to be_versioned }
 
   describe "#update!", versioning: true do
-    context "updating a skipped attribute" do
+    context "when updating a skipped attribute" do
       let(:t1) { Time.zone.local(2015, 7, 15, 20, 34, 0) }
       let(:t2) { Time.zone.local(2015, 7, 15, 20, 34, 30) }
 
@@ -16,6 +16,48 @@ RSpec.describe Skipper, type: :model, versioning: true do
           skipper.update!(another_timestamp: t2)
         }.not_to(change { skipper.versions.length })
       end
+    end
+  end
+
+  describe "#touch" do
+    let(:t1) { Time.zone.local(2015, 7, 15, 20, 34, 0) }
+    let(:t2) { Time.zone.local(2015, 7, 15, 20, 34, 30) }
+
+    if ActiveRecord.gem_version >= Gem::Version.new("6")
+      it "does not create a version for skipped attributes" do
+        skipper = Skipper.create!(another_timestamp: t1)
+        expect {
+          skipper.touch(:another_timestamp, time: t2)
+        }.not_to(change { skipper.versions.length })
+      end
+
+      it "does not create a version for ignored attributes" do
+        skipper = Skipper.create!(created_at: t1)
+        expect {
+          skipper.touch(:created_at, time: t2)
+        }.not_to(change { skipper.versions.length })
+      end
+    else
+      it "creates a version even for skipped attributes" do
+        skipper = Skipper.create!(another_timestamp: t1)
+        expect {
+          skipper.touch(:another_timestamp, time: t2)
+        }.to(change { skipper.versions.length })
+      end
+
+      it "creates a version even for ignored attributes" do
+        skipper = Skipper.create!(created_at: t1)
+        expect {
+          skipper.touch(:created_at, time: t2)
+        }.to(change { skipper.versions.length })
+      end
+    end
+
+    it "creates a version for non-skipped timestamps" do
+      skipper = Skipper.create!
+      expect {
+        skipper.touch
+      }.to(change { skipper.versions.length })
     end
   end