+++ spec/models/song_spec.rb
@@ -14,4 +14,25 @@ require "spec_helper"
       expect(result.event).to eq("create")
     end
   end
+
+  context "when the default accessor, length=, is overwritten" do
+    it "returns overwritten value on reified instance" do
+      song = Song.create(length: 4)
+      song.update(length: 5)
+      expect(song.length).to(eq(5))
+      expect(song.versions.last.reify.length).to(eq(4))
+    end
+  end
+
+  context "when song name is a virtual attribute (no such db column)" do
+    it "returns overwritten virtual attribute on the reified instance" do
+      song = Song.create(length: 4)
+      song.update(length: 5)
+      song.name = "Good Vibrations"
+      song.save
+      song.name = "Yellow Submarine"
+      expect(song.name).to(eq("Yellow Submarine"))
+      expect(song.versions.last.reify.name).to(eq("Good Vibrations"))
+    end
+  end
 end