+++ spec/requests/articles_spec.rb
@@ -5,7 +5,7 @@ require "spec_helper"
 RSpec.describe "Articles management", type: :request, order: :defined do
   let(:valid_params) { { article: { title: "Doh", content: FFaker::Lorem.sentence } } }
 
-  context "versioning disabled" do
+  context "with versioning disabled" do
     specify { expect(PaperTrail).not_to be_enabled }
 
     it "does not create a version" do
@@ -19,7 +19,7 @@ RSpec.describe "Articles management", type: :request, order: :defined do
   with_versioning do
     let(:article) { Article.last }
 
-    context "`current_user` method returns a `String`" do
+    context "when `current_user` method returns a `String`" do
       it "sets that value as the `whodunnit`" do
         expect {
           post articles_path, params: valid_params